Refactor: Move common stuff to common
