+++ src/components/ItemCard/ResourceCardContent.tsx
@@ -2,10 +2,10 @@ import React, { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 
+import GaiaXButton from '../../common/components/buttons/GaiaXButton';
 import Markdown from '../../common/markdown/Markdown';
 import { Resource } from '../../types/resources.model';
 import Title from '../Title/Title';
-import GaiaXButton from '../buttons/GaiaXButton';
 
 import styles from './ItemCard.module.css';
