+++ src/components/modals/LanguageModal.tsx
@@ -1,4 +1,4 @@
-import GaiaXButton from 'components/buttons/GaiaXButton';
+import GaiaXButton from 'common/components/buttons/GaiaXButton';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
