+++ src/components/ontologies/Ontologies.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
+import Horizontal from '../../common/components/./layouts/Horizontal';
+import Main from '../../common/components/./layouts/Main';
+import Vertical from '../../common/components/./layouts/Vertical';
+import ShowMapButton from '../../common/components/buttons/ShowMapButton';
 import RDFVisualization from '../../utils/RDFVisualization';
 import ItemCard from '../ItemCard/ItemCard';
-import ShowMapButton from '../buttons/ShowMapButton';
 import CardContainer from '../cards/CardContainer';
 import Header from '../header/Header';
-import Horizontal from '../layout/Horizontal';
-import Main from '../layout/Main';
-import Vertical from '../layout/Vertical';
 import LoadingIndicator from '../loading_view/LoadingIndicator';
 import NoContent from '../nocontent/NoContent';
 import SearchBar from '../searchBar/SearchBar';
